SEP1/web-app/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
milvus
openai
python-dotenv
werkzeug
langchain
langchain_openai
langchain_community
requests
faiss-cpu
chromadb